day18/main_test.go
package main

import (
	"testing"
)

var example = `5,4
4,2
4,5
3,0
2,1
6,3
2,4
1,5
0,6
3,3
2,6
5,1
1,2
5,5
2,5
6,5
1,4
0,4
6,4
1,1
6,1
1,0
0,5
1,6
2,0`

func Test_part1(t *testing.T) {
	tests := []struct {
		name  string
		input string
		want  int
		dim   int
		count int
	}{
		{
			name:  "example",
			input: example,
			want:  22,
			dim:   6,
			count: 12,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part1(tt.input, tt.dim, tt.count); got != tt.want {
				t.Errorf("part1() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_part2(t *testing.T) {
	tests := []struct {
		name  string
		input string
		want  string
		dim   int
	}{
		{
			name:  "example",
			input: example,
			want:  "6,1",
			dim:   6,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part2(tt.input, tt.dim); got != tt.want {
				t.Errorf("part2() = %v, want %v", got, tt.want)
			}
		})
	}
}
